# Speed up Color.random() via bit math