@pushfoo 2024-02-26
TL;DR of test failures: we made asssumptions when monkeypatching, working it out via discord